backend/pyproject.toml
[tool.poetry]
name = "voice-ai-backend"
version = "1.0.0"
description = "Voice AI Platform backend services"
authors = ["Your Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
deepgram-sdk = "^3.2.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
msgpack = "^1.0.7"
qdrant-client = "^1.7.0"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
celery = {extras = ["redis"], version = "^5.3.6"}
httpx = "^0.26.0"
pydantic = "^2.5.3"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
prometheus-client = "^0.19.0"
circuitbreaker = "^2.0.0"
azure-cognitiveservices-speech = "^1.35.0"
elevenlabs = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.3"
black = "^24.1.0"
ruff = "^0.2.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
